# Visualizer: switch to erase_if to improve performance